fix: add func for git origin (#931)
* fix: add func for git origin

* move test; deadcode

* timestamp
