+++ testdata/snowflake_provider_yaml/terraform/global/fogg.tf
@@ -106,16 +106,15 @@ variable "owner" {
 variable "tags" {
   type = object({ project : string, env : string, service : string, owner : string, managedBy : string })
   default = {
-    project               = "foo"
-    env                   = ""
-    service               = "global"
-    owner                 = "foo@example.com"
-    date                  = "2023-09-13"
-    terraformWorkspaceDir = "/terraform/global"
-    foggUser              = ""
-    gitRepository         = "git@github.com:chanzuckerberg/fogg"
-    commitHash            = ""
-    managedBy             = "terraform"
+    project                = "foo"
+    env                    = ""
+    service                = "global"
+    owner                  = "foo@example.com"
+    terraformLastApplyTime = timestamp()
+    terraformWorkspaceDir  = "/terraform/global"
+    foggUser               = ""
+    gitRepository          = "git@github.com:chanzuckerberg/fogg"
+    managedBy              = "terraform"
   }
 }
 # tflint-ignore: terraform_unused_declarations
