+++ testdata/v2_full_yaml/terraform/envs/staging/vpc/fogg.tf
@@ -117,16 +117,15 @@ variable "owner" {
 variable "tags" {
   type = object({ project : string, env : string, service : string, owner : string, managedBy : string })
   default = {
-    project               = "proj"
-    env                   = "staging"
-    service               = "vpc"
-    owner                 = "foo@example.com"
-    date                  = "2023-09-13"
-    terraformWorkspaceDir = "/terraform/envs/staging/vpc"
-    foggUser              = ""
-    gitRepository         = "git@github.com:chanzuckerberg/fogg"
-    commitHash            = ""
-    managedBy             = "terraform"
+    project                = "proj"
+    env                    = "staging"
+    service                = "vpc"
+    owner                  = "foo@example.com"
+    terraformLastApplyTime = timestamp()
+    terraformWorkspaceDir  = "/terraform/envs/staging/vpc"
+    foggUser               = ""
+    gitRepository          = "git@github.com:chanzuckerberg/fogg"
+    managedBy              = "terraform"
   }
 }
 # tflint-ignore: terraform_unused_declarations
