+++ testdata/v2_no_aws_provider_yaml/terraform/accounts/bar/fogg.tf
@@ -96,16 +96,15 @@ variable "owner" {
 variable "tags" {
   type = object({ project : string, env : string, service : string, owner : string, managedBy : string })
   default = {
-    project               = "proj"
-    env                   = "accounts"
-    service               = "bar"
-    owner                 = "foo@example.com"
-    date                  = "2023-09-13"
-    terraformWorkspaceDir = "/terraform/accounts/bar"
-    foggUser              = ""
-    gitRepository         = "git@github.com:chanzuckerberg/fogg"
-    commitHash            = ""
-    managedBy             = "terraform"
+    project                = "proj"
+    env                    = "accounts"
+    service                = "bar"
+    owner                  = "foo@example.com"
+    terraformLastApplyTime = timestamp()
+    terraformWorkspaceDir  = "/terraform/accounts/bar"
+    foggUser               = ""
+    gitRepository          = "git@github.com:chanzuckerberg/fogg"
+    managedBy              = "terraform"
   }
 }
 # tflint-ignore: terraform_unused_declarations
